New feature
It's now possible to pass a new parameter to weekly. The parameter is called `selectableDates`.

If the value of the parameter is an Array, then weekly will match the date against that array and, if found, it will allow the event to be created.

The value can also be a function, in which case the function will receive the date as a parameter and returning `true` will allow the event to be added whereas `false` won't.

Fixes #29